COMMERCE-579 Move the logic to LiferayConnectionProperties and use the reference connection properties if possible
