+++ modules/etl/talend/talend-definition/src/main/java/com/liferay/talend/connection/LiferayConnectionProperties.java
@@ -15,25 +15,35 @@
 package com.liferay.talend.connection;
 
 import com.liferay.talend.LiferayBaseComponentDefinition;
+import com.liferay.talend.exception.ExceptionUtils;
 import com.liferay.talend.runtime.LiferaySourceOrSinkRuntime;
 import com.liferay.talend.tliferayconnection.TLiferayConnectionDefinition;
 import com.liferay.talend.utils.PropertiesUtils;
 
+import java.io.IOException;
+
 import java.util.EnumSet;
+import java.util.List;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import org.talend.components.api.properties.ComponentPropertiesImpl;
 import org.talend.components.api.properties.ComponentReferenceProperties;
+import org.talend.daikon.NamedThing;
+import org.talend.daikon.i18n.GlobalI18N;
+import org.talend.daikon.i18n.I18nMessageProvider;
+import org.talend.daikon.i18n.I18nMessages;
 import org.talend.daikon.properties.PresentationItem;
 import org.talend.daikon.properties.Properties;
 import org.talend.daikon.properties.ValidationResult;
 import org.talend.daikon.properties.ValidationResult.Result;
+import org.talend.daikon.properties.ValidationResultMutable;
 import org.talend.daikon.properties.presentation.Form;
 import org.talend.daikon.properties.presentation.Widget;
 import org.talend.daikon.properties.property.Property;
 import org.talend.daikon.properties.property.PropertyFactory;
+import org.talend.daikon.properties.property.StringProperty;
 import org.talend.daikon.properties.service.Repository;
 import org.talend.daikon.sandbox.SandboxedInstance;
 
@@ -84,6 +94,111 @@ public class LiferayConnectionProperties
 	public void afterReferencedComponent() {
 		refreshLayout(getForm(Form.MAIN));
 		refreshLayout(getForm(Form.REFERENCE));
+	}
+
+	public void afterSiteFilter() {
+		refreshLayout(getForm(Form.MAIN));
+		refreshLayout(getForm(Form.REFERENCE));
+	}
+
+	public ValidationResult afterWebSiteURL() {
+		if (_log.isDebugEnabled()) {
+			_log.debug("Website URL: " + webSiteURL.getValue());
+		}
+
+		ValidationResultMutable validationResultMutable =
+			new ValidationResultMutable();
+
+		validationResultMutable.setStatus(Result.OK);
+
+		try (SandboxedInstance sandboxedInstance =
+				LiferayBaseComponentDefinition.getSandboxedInstance(
+					LiferayBaseComponentDefinition.
+						RUNTIME_SOURCE_OR_SINK_CLASS_NAME)) {
+
+			LiferaySourceOrSinkRuntime liferaySourceOrSinkRuntime =
+				(LiferaySourceOrSinkRuntime)sandboxedInstance.getInstance();
+
+			liferaySourceOrSinkRuntime.initialize(
+				null, getReferencedConnectionProperties());
+
+			ValidationResult validationResult =
+				liferaySourceOrSinkRuntime.validate(null);
+
+			validationResultMutable.setMessage(validationResult.getMessage());
+			validationResultMutable.setStatus(validationResult.getStatus());
+
+			if (validationResultMutable.getStatus() ==
+					ValidationResult.Result.OK) {
+
+				try {
+					webSite.setValue(
+						liferaySourceOrSinkRuntime.getActualWebSiteName(
+							webSiteURL.getValue()));
+				}
+				catch (IOException ioe) {
+					validationResult =
+						ExceptionUtils.exceptionToValidationResult(ioe);
+
+					validationResultMutable.setMessage(
+						validationResult.getMessage());
+					validationResultMutable.setStatus(
+						validationResult.getStatus());
+				}
+			}
+		}
+
+		if (validationResultMutable.getStatus() ==
+				ValidationResult.Result.ERROR) {
+		}
+
+		refreshLayout(getForm(Form.MAIN));
+		refreshLayout(getForm(Form.REFERENCE));
+
+		return validationResultMutable;
+	}
+
+	public ValidationResult beforeWebSiteURL() {
+		try (SandboxedInstance sandboxedInstance =
+				LiferayBaseComponentDefinition.getSandboxedInstance(
+					LiferayBaseComponentDefinition.
+						RUNTIME_SOURCE_OR_SINK_CLASS_NAME)) {
+
+			LiferaySourceOrSinkRuntime liferaySourceOrSinkRuntime =
+				(LiferaySourceOrSinkRuntime)sandboxedInstance.getInstance();
+
+			liferaySourceOrSinkRuntime.initialize(
+				null, getReferencedConnectionProperties());
+
+			ValidationResultMutable validationResultMutable =
+				new ValidationResultMutable();
+
+			ValidationResult validationResult =
+				liferaySourceOrSinkRuntime.validate(null);
+
+			validationResultMutable.setStatus(validationResult.getStatus());
+
+			if (validationResultMutable.getStatus() == Result.OK) {
+				try {
+					List<NamedThing> webSites =
+						liferaySourceOrSinkRuntime.getAvailableWebSites();
+
+					if (webSites.isEmpty()) {
+						validationResultMutable.setMessage(
+							i18nMessages.getMessage(
+								"error.validation.websites"));
+						validationResultMutable.setStatus(Result.ERROR);
+					}
+
+					webSiteURL.setPossibleNamedThingValues(webSites);
+				}
+				catch (Exception e) {
+					return ExceptionUtils.exceptionToValidationResult(e);
+				}
+			}
+
+			return validationResultMutable;
+		}
 	}
 
 	@Override
@@ -198,6 +313,22 @@ public class LiferayConnectionProperties
 
 		mainForm.addRow(anonymousLogin);
 
+		mainForm.addRow(siteFilter);
+
+		Widget webSiteURLWidget = Widget.widget(webSiteURL);
+
+		webSiteURLWidget.setCallAfter(true);
+		webSiteURLWidget.setWidgetType(
+			Widget.NAME_SELECTION_REFERENCE_WIDGET_TYPE);
+
+		mainForm.addRow(webSiteURLWidget);
+
+		Widget webSiteWidget = Widget.widget(webSite);
+
+		webSiteWidget.setReadonly(true);
+
+		mainForm.addColumn(webSiteWidget);
+
 		// A form for a reference to a connection, used in a tLiferayInput
 		// for example
 
@@ -210,7 +341,36 @@ public class LiferayConnectionProperties
 
 		referenceForm.addRow(referencedComponentWidget);
 
-		referenceForm.addRow(mainForm);
+		Widget loginReferenceWidget = Widget.widget(loginType);
+
+		loginReferenceWidget.setWidgetType(Widget.ENUMERATION_WIDGET_TYPE);
+
+		referenceForm.addRow(loginReferenceWidget);
+
+		referenceForm.addRow(endpoint);
+
+		referenceForm.addRow(userId);
+
+		referenceForm.addRow(password);
+
+		referenceForm.addRow(anonymousLogin);
+
+		referenceForm.addRow(siteFilter);
+
+		Widget webSiteURLReferenceWidget = Widget.widget(webSiteURL);
+
+		webSiteURLReferenceWidget.setCallAfter(true);
+		webSiteURLReferenceWidget.setLongRunning(true);
+		webSiteURLReferenceWidget.setWidgetType(
+			Widget.NAME_SELECTION_REFERENCE_WIDGET_TYPE);
+
+		referenceForm.addRow(webSiteURLReferenceWidget);
+
+		Widget webSiteReferenceWidget = Widget.widget(webSite);
+
+		webSiteReferenceWidget.setReadonly(true);
+
+		referenceForm.addColumn(webSiteReferenceWidget);
 
 		refreshLayout(referenceForm);
 
@@ -238,7 +398,10 @@ public class LiferayConnectionProperties
 		forceHttps.setValue(false);
 		loginType.setValue(LoginType.Basic);
 		password.setValue(_PASSWORD);
+		siteFilter.setValue(false);
 		userId.setValue(_USER_ID);
+		webSite.setValue("");
+		webSiteURL.setValue("");
 	}
 
 	public ValidationResult validateTestConnection() {
@@ -294,9 +457,13 @@ public class LiferayConnectionProperties
 	public ComponentReferenceProperties<LiferayConnectionProperties>
 		referencedComponent = new ComponentReferenceProperties<>(
 			"referencedComponent", TLiferayConnectionDefinition.COMPONENT_NAME);
+	public Property<Boolean> siteFilter = PropertyFactory.newBoolean(
+		"siteFilter");
 	public PresentationItem testConnection = new PresentationItem(
 		"testConnection");
 	public Property<String> userId = PropertyFactory.newString("userId");
+	public Property<String> webSite = PropertyFactory.newString("webSite");
+	public StringProperty webSiteURL = PropertyFactory.newString("webSiteURL");
 
 	public enum LoginType {
 
@@ -317,6 +484,16 @@ public class LiferayConnectionProperties
 	protected SandboxedInstance getRuntimeSandboxedInstance() {
 		return LiferayBaseComponentDefinition.getSandboxedInstance(
 			LiferayBaseComponentDefinition.RUNTIME_SOURCE_OR_SINK_CLASS_NAME);
+	}
+
+	protected static final I18nMessages i18nMessages;
+
+	static {
+		I18nMessageProvider i18nMessageProvider =
+			GlobalI18N.getI18nMessageProvider();
+
+		i18nMessages = i18nMessageProvider.getI18nMessages(
+			LiferayConnectionProperties.class);
 	}
 
 	/**
