+++ modules/etl/talend/talend-definition/src/main/java/com/liferay/talend/resource/LiferayResourceProperties.java
@@ -19,7 +19,6 @@ import com.liferay.talend.connection.LiferayConnectionProperties;
 import com.liferay.talend.connection.LiferayConnectionPropertiesProvider;
 import com.liferay.talend.exception.ExceptionUtils;
 import com.liferay.talend.runtime.LiferaySourceOrSinkRuntime;
-import com.liferay.talend.utils.PropertiesUtils;
 import com.liferay.talend.utils.URIUtils;
 
 import java.io.IOException;
@@ -139,77 +138,6 @@ public class LiferayResourceProperties
 		return validationResultMutable;
 	}
 
-	public void afterSiteFilter() {
-		resource.setValue("");
-		resourceURL.setValue("");
-
-		refreshLayout(getForm(Form.MAIN));
-		refreshLayout(getForm(Form.REFERENCE));
-	}
-
-	public ValidationResult afterWebSiteURL() {
-		resource.setValue("");
-		resourceURL.setValue("");
-
-		if (_log.isDebugEnabled()) {
-			_log.debug("Website URL: " + webSiteURL.getValue());
-		}
-
-		ValidationResultMutable validationResultMutable =
-			new ValidationResultMutable();
-
-		validationResultMutable.setStatus(Result.OK);
-
-		try (SandboxedInstance sandboxedInstance =
-				LiferayBaseComponentDefinition.getSandboxedInstance(
-					LiferayBaseComponentDefinition.
-						RUNTIME_SOURCE_OR_SINK_CLASS_NAME)) {
-
-			LiferaySourceOrSinkRuntime liferaySourceOrSinkRuntime =
-				(LiferaySourceOrSinkRuntime)sandboxedInstance.getInstance();
-
-			liferaySourceOrSinkRuntime.initialize(
-				null, getEffectiveLiferayConnectionProperties());
-
-			ValidationResult validationResult =
-				liferaySourceOrSinkRuntime.validate(null);
-
-			validationResultMutable.setMessage(validationResult.getMessage());
-			validationResultMutable.setStatus(validationResult.getStatus());
-
-			if (validationResultMutable.getStatus() ==
-					ValidationResult.Result.OK) {
-
-				try {
-					webSite.setValue(
-						liferaySourceOrSinkRuntime.getActualWebSiteName(
-							webSiteURL.getValue()));
-				}
-				catch (IOException ioe) {
-					validationResult =
-						ExceptionUtils.exceptionToValidationResult(ioe);
-
-					validationResultMutable.setMessage(
-						validationResult.getMessage());
-					validationResultMutable.setStatus(
-						validationResult.getStatus());
-				}
-			}
-		}
-
-		if (validationResultMutable.getStatus() ==
-				ValidationResult.Result.ERROR) {
-
-			resource.setValue("");
-			resourceURL.setValue("");
-		}
-
-		refreshLayout(getForm(Form.MAIN));
-		refreshLayout(getForm(Form.REFERENCE));
-
-		return validationResultMutable;
-	}
-
 	public ValidationResult beforeResourceURL() throws Exception {
 		try (SandboxedInstance sandboxedInstance =
 				LiferayBaseComponentDefinition.getSandboxedInstance(
@@ -230,14 +158,18 @@ public class LiferayResourceProperties
 
 			validationResultMutable.setStatus(validationResult.getStatus());
 
+			LiferayConnectionProperties liferayConnectionProperties =
+				getEffectiveLiferayConnectionProperties();
+
 			if (validationResultMutable.getStatus() == Result.OK) {
 				try {
 					List<NamedThing> resourceNames = null;
 
-					if (siteFilter.getValue()) {
+					if (liferayConnectionProperties.siteFilter.getValue()) {
 						resourceNames =
 							liferaySourceOrSinkRuntime.getResourceList(
-								webSiteURL.getValue());
+								liferayConnectionProperties.webSiteURL.
+									getValue());
 					}
 					else {
 						resourceNames =
@@ -262,49 +194,6 @@ public class LiferayResourceProperties
 		}
 	}
 
-	public ValidationResult beforeWebSiteURL() {
-		try (SandboxedInstance sandboxedInstance =
-				LiferayBaseComponentDefinition.getSandboxedInstance(
-					LiferayBaseComponentDefinition.
-						RUNTIME_SOURCE_OR_SINK_CLASS_NAME)) {
-
-			LiferaySourceOrSinkRuntime liferaySourceOrSinkRuntime =
-				(LiferaySourceOrSinkRuntime)sandboxedInstance.getInstance();
-
-			liferaySourceOrSinkRuntime.initialize(
-				null, getEffectiveLiferayConnectionProperties());
-
-			ValidationResultMutable validationResultMutable =
-				new ValidationResultMutable();
-
-			ValidationResult validationResult =
-				liferaySourceOrSinkRuntime.validate(null);
-
-			validationResultMutable.setStatus(validationResult.getStatus());
-
-			if (validationResultMutable.getStatus() == Result.OK) {
-				try {
-					List<NamedThing> webSites =
-						liferaySourceOrSinkRuntime.getAvailableWebSites();
-
-					if (webSites.isEmpty()) {
-						validationResultMutable.setMessage(
-							i18nMessages.getMessage(
-								"error.validation.websites"));
-						validationResultMutable.setStatus(Result.ERROR);
-					}
-
-					webSiteURL.setPossibleNamedThingValues(webSites);
-				}
-				catch (Exception e) {
-					return ExceptionUtils.exceptionToValidationResult(e);
-				}
-			}
-
-			return validationResultMutable;
-		}
-	}
-
 	@Override
 	public LiferayConnectionProperties getLiferayConnectionProperties() {
 		return connection;
@@ -313,13 +202,6 @@ public class LiferayResourceProperties
 	@Override
 	public void refreshLayout(Form form) {
 		super.refreshLayout(form);
-
-		String formName = form.getName();
-
-		if (formName.equals(Form.MAIN) || formName.equals(Form.REFERENCE)) {
-			PropertiesUtils.setHidden(form, webSite, !siteFilter.getValue());
-			PropertiesUtils.setHidden(form, webSiteURL, !siteFilter.getValue());
-		}
 	}
 
 	public void setSchemaListener(ISchemaListener schemaListener) {
@@ -347,9 +229,6 @@ public class LiferayResourceProperties
 		condition.setValue("");
 		resource.setValue("");
 		resourceURL.setValue("");
-		siteFilter.setValue(false);
-		webSite.setValue("");
-		webSiteURL.setValue("");
 	}
 
 	public ValidationResult validateValidateCondition() {
@@ -358,7 +237,10 @@ public class LiferayResourceProperties
 
 		validationResultMutable.setStatus(Result.OK);
 
-		String endpointUrl = connection.endpoint.getValue();
+		LiferayConnectionProperties liferayConnectionProperties =
+			getEffectiveLiferayConnectionProperties();
+
+		String endpointUrl = liferayConnectionProperties.endpoint.getValue();
 
 		try {
 			URIUtils.addQueryConditionToURL(endpointUrl, condition.getValue());
@@ -389,12 +271,8 @@ public class LiferayResourceProperties
 	public StringProperty resourceURL = PropertyFactory.newString(
 		"resourceURL");
 	public ISchemaListener schemaListener;
-	public Property<Boolean> siteFilter = PropertyFactory.newBoolean(
-		"siteFilter");
 	public transient PresentationItem validateCondition = new PresentationItem(
 		"validateCondition");
-	public Property<String> webSite = PropertyFactory.newString("webSite");
-	public StringProperty webSiteURL = PropertyFactory.newString("webSiteURL");
 
 	protected LiferayConnectionProperties
 		getEffectiveLiferayConnectionProperties() {
@@ -448,21 +326,6 @@ public class LiferayResourceProperties
 	private void _setupMainForm() {
 		Form resourceSelectionForm = Form.create(this, Form.MAIN);
 
-		resourceSelectionForm.addRow(siteFilter);
-
-		Widget webSiteURLWidget = Widget.widget(webSiteURL);
-
-		webSiteURLWidget.setCallAfter(true);
-		webSiteURLWidget.setWidgetType(Widget.NAME_SELECTION_AREA_WIDGET_TYPE);
-
-		resourceSelectionForm.addRow(webSiteURLWidget);
-
-		Widget webSiteWidget = Widget.widget(webSite);
-
-		webSiteWidget.setReadonly(true);
-
-		resourceSelectionForm.addColumn(webSiteWidget);
-
 		Widget resourceURLWidget = Widget.widget(resourceURL);
 
 		resourceURLWidget.setCallAfter(true);
@@ -490,23 +353,6 @@ public class LiferayResourceProperties
 
 	private void _setupReferenceForm() {
 		Form referenceForm = Form.create(this, Form.REFERENCE);
-
-		referenceForm.addRow(siteFilter);
-
-		Widget webSiteURLReferenceWidget = Widget.widget(webSiteURL);
-
-		webSiteURLReferenceWidget.setCallAfter(true);
-		webSiteURLReferenceWidget.setLongRunning(true);
-		webSiteURLReferenceWidget.setWidgetType(
-			Widget.NAME_SELECTION_REFERENCE_WIDGET_TYPE);
-
-		referenceForm.addRow(webSiteURLReferenceWidget);
-
-		Widget webSiteReferenceWidget = Widget.widget(webSite);
-
-		webSiteReferenceWidget.setReadonly(true);
-
-		referenceForm.addColumn(webSiteReferenceWidget);
 
 		Widget resourceURLReferenceWidget = Widget.widget(resourceURL);
