+++ modules/etl/talend/talend-definition/src/main/resources/com/liferay/talend/connection/LiferayConnectionProperties.properties
@@ -1,3 +1,5 @@
+error.validation.websites=Unable to find Website resources. Please check the endpoint URL specified for the server.
+
 form.Advanced.displayName=Advanced
 form.Advanced.title=Advanced
 form.Main.displayName=Main
@@ -19,4 +21,7 @@ property.name.displayName=Name
 property.password.displayName=Password
 property.readTimeout.displayName=REST Read Timeout (s)
 property.testConnection.displayName=Test Connection
+property.siteFilter.displayName=Site Filter
 property.userId.displayName=User
+property.webSite.displayName=Website Name
+property.webSiteURL.displayName=Website URL
