+++ modules/etl/talend/talend-definition/src/main/resources/com/liferay/talend/resource/LiferayResourceProperties.properties
@@ -1,6 +1,5 @@
 error.validation.resources=Unable to find any exposed resources. Please check the endpoint URL specified for the server.
 error.validation.resourceType=Unable to determine the resource''s type. The resource must have at least one entity.
-error.validation.websites=Unable to find Website resources. Please check the endpoint URL specified for the server.
 
 form.Advanced.displayName=Advanced
 form.Advanced.title=Advanced
@@ -11,8 +10,5 @@ form.Reference.title=Reference
 
 property.condition.displayName=Condition
 property.resource.displayName=Resource Type
-property.resourceURL.displayName=Resource URL
-property.siteFilter.displayName=Site Filter
 property.validateCondition.displayName=Validate Condition
-property.webSite.displayName=Website Name
-property.webSiteURL.displayName=Website URL
+property.resourceURL.displayName=Resource URL
